requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
pymysql==1.1.1
pydantic==2.9.2
pydantic-settings==2.5.2
pyjwt==2.9.0
bcrypt==4.2.0
python-multipart==0.0.12
